requirements.txt
Flask==3.0.3
psycopg[binary]==3.2.3
python-dotenv==1.0.1
gunicorn==22.0.0
